requirements.txt
blinker==1.9.0
build==1.2.2.post1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.2.5
packaging==25.0
pip-tools==7.4.1
psutil==7.0.0
pyproject_hooks==1.2.0
python-dotenv==1.1.0
scikit-learn==1.6.1
scipy==1.15.2
setuptools==79.0.1
threadpoolctl==3.6.0
Werkzeug==3.1.3
wheel==0.45.1
